moved the env config from common to core.
