+++ framework/core/src/environment/getEnvVar.ts
