+++ framework/core/src/http/telemetry/recordMetric.ts
@@ -1,4 +1,3 @@
-import { getEnvVar } from '@forklaunch/common';
 import { AnySchemaValidator } from '@forklaunch/validator';
 import {
   ATTR_HTTP_REQUEST_METHOD,
@@ -13,6 +12,7 @@ import {
   ParamsDictionary,
   VersionedRequests
 } from '..';
+import { getEnvVar } from '../../environment';
 import { ATTR_API_NAME, ATTR_CORRELATION_ID } from './constants';
 import { httpRequestsTotalCounter } from './openTelemetryCollector';
